requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-admin-interface==0.19.1
django-colorfield==0.7.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
yarl==1.7.2
gunicorn==20.1.0
